Implement fully unrolled ec-point-double module
This is written in a similar style barrett reduction, having a
stage-by-stage breakdown to individual modules. While this is not
necessarily the most efficient in Flipflop usage, it is simple to understand,
and our designs probably won't be found be FF resources anwyay.

Testing:
- Using in karatsuba-ofman + barrett reduction as the underlying
  multiplier
- Using in montgomery multiplication

In the case of montgomery multiplication, the test code has to convert
test the circuit output into montgomery space and verify that.