Add Config.t parameter for Karatsuba_ofman_mult (#3)
Replace single [depth] parameter with a more flexible [Config.t] to control various parameters of karatsuba ofman multiplier. This knobs includes: the ground multiplier latency, the type of ground multiplier and the number of pipeline stages in the final adder. More knobs can be added to the Config.t record type as needed.