Add a pipelined adder component (#4)
This pipelined adder component is similar to the pipelined modulo adder component, except it supports more than 2 arguments and doesn't perform any modulo operation. See adder_pipe.mli for comments about the generated pipeline adder's architecture.

To test this implementation

This is immediately used in the montgomery multiplier, replacing the modulo add component with this newly created regular add component. The existing tests with the montgomery multiplication still works
Tests to formally prove that the summation is equivalent to a naive summation implementation using Hardcaml_verify